LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbrequest_result-expected.txt


PASS IDBRequest.result throws if ready state is pending
